==> Notebooks/02_Event_Study_Returns.ipynb <==
2. Event Study Metrics Calculations

This notebook computes abnormal returns (AR), cumulative abnormal returns (CAR), and other event-study metrics for each IPO in the sample.


import pandas as pd
import os
sample1 = pd.read_csv("clean_sample1.csv")
sample2 = pd.read_csv("clean_sample2.csv")
nifty = pd.read_csv("clean_nifty_returns.csv")

sample1.head()
sample2.head()
nifty.head()

import pandas as pd
import os
sample1 = pd.read_csv("clean_sample1.csv")
sample2 = pd.read_csv("clean_sample2.csv")
nifty = pd.read_csv("clean_nifty_returns.csv")

#Removing the first NaN row because pct_change will naturally create one NaN
sample1 = sample1.dropna()
sample2 = sample2.dropna()
 

2.1 Calculation of Abnormal Return and Cumulative Abnormal Return

(a) Abnormal Return :  ARt​= RtIPO​−RtNIFTY​

(b) Cumulative Abnormal Return: CARt1​,t2​​=t=t1​∑t2​​ARt​

(c) Compute AR per IPO

(d) Compute CAR for the event window

(e) Validate results with summary checks

#Abnormal Return Calculation
sample1["abnormal_return"] = sample1["ipo_return"] - sample1["nifty_return"]
sample2["abnormal_return"] = sample2["ipo_return"] - sample2["nifty_return"]

#CAR Calculation (Cumulative Abnormal Return)
sample1["car"] = sample1['abnormal_return'].cumsum()
sample2["car"] = sample2['abnormal_return'].cumsum()
 


2.2 Calculation of All Metrics

This section includes calculation of various relevant metrics computed via formulas implemented directly in Python

def compute_30_day_metrics(df):
    df_30 = df.head(30)             #First 30 days after listing

    mean_return = df_30['ipo_return'].mean()
    mean_abnormal_30 = df_30["abnormal_return"].mean()
    car_30 = df_30["car"].iloc[-1]
    volatility_return_30 = df_30['ipo_return'].std()
    volatility_abnormal = df_30["abnormal_return"].std()

    #beta calculation
    beta = df_30['ipo_return'].cov(df_30['nifty_return']) / df_30['nifty_return'].var()

    #Sharpe Like Metric but without Risk-free rate
    sharpe_like = mean_return/volatility_return_30 if volatility_return_30 != 0 else None

    #Max_Drawdown
    cum_curve = (1 + df_30['ipo_return']).cumprod()
    running_max = cum_curve.cummax()
    drawdown = (cum_curve - running_max) / running_max
    max_drawdown = drawdown.min()

    return {
    "mean_return": mean_return,
    "mean_abnormal_30": mean_abnormal_30,
    "car_30": car_30,
    "volatility_return_30": volatility_return_30,
    "volatility_abnormal": volatility_abnormal,
    "beta": beta,
    "sharpe_like": sharpe_like,
    "max_drawdown": max_drawdown
}


sample1_metrics = compute_30_day_metrics(sample1)
sample2_metrics = compute_30_day_metrics(sample2)
print (sample1_metrics)
print (sample2_metrics)

2.2 Comparision of Metrics

#COMPARISIONS
sample1_summary = {
    'mean_return': 0.01646627623624749,
    'mean_abnormal_30': 0.015630431752341144,
    'car_30': 0.14067388577107032,
    'volatility_return_30': 0.018059501648673135,
    'volatility_abnormal': 0.017204581295557363,
    'beta': 17.435858126734892,
    'sharpe_like': 0.911779104239972,
    'max_drawdown': -0.01769911504424772
}

sample2_summary = {
    'mean_return': 0.012090668953962522,
    'mean_abnormal_30': 0.011254824470056178,
    'car_30': 0.10129342023050561,
    'volatility_return_30': 0.011643540248923749,
    'volatility_abnormal': 0.011439452894603411,
    'beta': 3.1464102771212685,
    'sharpe_like': 1.0384014393800978,
    'max_drawdown': -0.008888888888888839
}


import pandas as pd
comparision = pd.DataFrame({ "Sample_1": sample1_summary, "Sample_2": sample2_summary})
comparision

#With a Difference column
comparision['Difference(S1 - S2)'] = comparision["Sample_1"] - comparision["Sample_2"]
comparision

 


2.3 Exporting Final Files

# Save final processed DataFrames to the existing Processed_Data folder
output_path_sample1 = "../Data/Processed_Data/final_sample1.csv"
output_path_sample2 = "../Data/Processed_Data/final_sample2.csv"

sample1.to_csv(output_path_sample1, index=False)
sample2.to_csv(output_path_sample2, index=False)
